adk-web/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.4.0",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
]

[tool.uv.workspace]
members = [
    "1.agent_basic",
]
